refactor: rename provider defaults module to config
